+++ harmony_utils/src/main/ets/entity/constranint.ts
@@ -1,0 +1,5 @@
+import { ArrayList, HashMap } from "@kit.ArkTS";
+
+
+export type CommonSingleType = Object | String | Number | Boolean | null | undefined; //通用单个联合类型
+export type CommonAllType = CommonSingleType | Array<CommonSingleType> | ArrayList<CommonSingleType> | HashMap<CommonSingleType, CommonSingleType>; //通用所有联合类型